design: findMakerService 삭제 후 userService에 코드 머지
