+++ src/pages/finding-maker.tsx
@@ -5,7 +5,7 @@ import CardFindMaker from "@/components/Common/CardFindMaker";
 import SearchBar from "@/components/Common/SearchBar";
 import Link from 'next/link';
 import useAuthStore from "@/stores/useAuthStore";
-import { getMakers, ServiceType } from '@/services/findMakerService';
+import { getMakers, ServiceType } from '@/services/userService';
 import { useInfiniteQuery } from "@tanstack/react-query";
 import { useInView } from "react-intersection-observer";
 import followService, { FollowedCardProps } from "@/services/followService";
