+++ src/services/userService.ts
@@ -1,7 +1,16 @@
 import { api } from "./api";
 
 type Role = "DREAMER" | "MAKER";
-
+export type ServiceType =
+  | "SHOPPING"
+  | "FOOD_TOUR"
+  | "ACTIVITY"
+  | "CULTURE"
+  | "FESTIVAL"
+  | "RELAXATION"
+  | "REQUEST"
+  | "PENDING"
+  | "CONFIRMED";
 export interface UserInfo {
   id: string;
   role: Role;
@@ -69,6 +78,27 @@ interface MakerReviewResponse {
 interface MakerReviewParams {
   page?: number;
   pageSize?: number;
+}
+
+export interface Maker {
+  id: string;
+  nickName: string;
+  description: string;
+  detailDescription: string;
+  image: string;
+  gallery: string;
+  averageRating: number;
+  totalReviews: number;
+  totalFollows: number;
+  totalConfirms: number;
+  serviceTypes: ServiceType[];
+  serviceArea: string[];
+  isFollowed: boolean;
+}
+
+interface MakerResponse {
+  totalCount: number;
+  list: Maker[];
 }
 
 const userService = {
@@ -164,3 +194,26 @@ const userService = {
 };
 
 export default userService;
+
+export const getMakers = async (
+  orderBy: string,
+  serviceArea: string,
+  serviceType: string,
+  pageParam?: number,
+  pageSize?: number,
+  keyword?: string,
+): Promise<MakerResponse> => {
+  try {
+    const url =
+      `/users/makers?page=${pageParam}&pageSize=${pageSize}` +
+      `${orderBy ? `&orderBy=${orderBy}` : ""}` +
+      `${serviceArea ? `&serviceArea=${serviceArea}` : ""}` +
+      `${serviceType ? `&serviceType=${serviceType}` : ""}` +
+      `${keyword ? `&keyword=${keyword}` : ""}`;
+    const response = await api.get(url);
+    return response as MakerResponse;
+  } catch (error) {
+    console.error("Error fetching makers:", error);
+    throw error;
+  }
+};
